refactor(index): have index just export what it needs to
