+++ test/redux-observable-spec.js
